Create SelectFileDirectory
To use this script, open the Excel VBA editor (press Alt+F11), create a new module (Insert > Module), and paste the code into the module. Then, run the script by clicking on the Run button or by pressing F5.

This script will display a File Dialog that allows you to select a folder. When you select a folder and click OK, the file path of the selected folder will be placed in cell A1 of the active worksheet.

Note that the Application.FileDialog method and the Range.Value property are specific to Microsoft Excel and may not be available in other spreadsheet applications. To make this script cross-platform, you will need to use a different method to select a file or folder and a different property to set the value of a cell.